backend/requirements.txt
Django==4.1.5
gunicorn==20.1.0
djangorestframework==3.14.0
psycopg2==2.9.5
django-cors-headers